CHG: Fix BasalforcingsLaddie analyses.
First, the Laddie simulation exhibits oscillations during sub-time step integration. To address this, we store each `BasalforcingsLaddieMeltingRate` value in `BasalforcingsDiagLaddieMeltingRate` and normalize it by dividing by the total number of sub-time steps in Laddie. After finishing Laddie simulation, `BasalforcingsDiagLaddieMeltingRate` is duplicated to `BasalforcingsFloatingiceMeltingRateEnum`.

The momentum term in the plume model is nonlinear due to the convection term. To handle this, we solve the momentum equation iteratively using the implementation in `solutionsequence_laddie_nonlinear.cpp`.

BasalforcingsLaddieHead and BasalforcingsLaddieSalt
* Correct stabilization schemes, including artificial diffusion and streamline upwinding methods.
* Fix the treatment of the advection term.

BasalforcingsLaddieMomentumAnalysis.cpp
* Correct the handling of the convection term: values are now assigned only to the diagonal components, excluding off-diagonal components.
* Move the Coriolis and drag (friction) terms from the stiffness matrix to the loading vector.